pyproject.toml
[project]
name = "fastapi-graphql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "strawberry-graphql[debug-server]>=0.266.0",
]
